relocate geometryScheme to finiteVolume
