+++ src/CMakeLists.txt
@@ -19,10 +19,10 @@ target_sources(
           "mesh/unstructured/boundaryMesh.cpp"
           "mesh/unstructured/communicator.cpp"
           "mesh/unstructured/unstructuredMesh.cpp"
-          "mesh/stencil/stencilDataBase.cpp"
-          "mesh/stencil/stencilDataBase.cpp"
-          "mesh/stencil/fvccGeometryScheme.cpp"
-          "mesh/stencil/basicFvccGeometryScheme.cpp"
+          "finiteVolume/cellCentred/stencil/stencilDataBase.cpp"
+          "finiteVolume/cellCentred/stencil/stencilDataBase.cpp"
+          "finiteVolume/cellCentred/stencil/geometryScheme.cpp"
+          "finiteVolume/cellCentred/stencil/basicGeometryScheme.cpp"
           "finiteVolume/cellCentred/boundary/boundary.cpp"
           "finiteVolume/cellCentred/operators/gaussGreenGrad.cpp"
           "finiteVolume/cellCentred/operators/gaussGreenDiv.cpp"
